Renamed symbtrdataextractor/offset.py -> symbtrdataextractor/OffsetProcessor.py
