+++ symbtrdataextractor/OffsetProcessor.py
